requirements.txt
python-telegram-bot==20.7
flask==2.3.3
pandas==1.5.3
openpyxl==3.1.2
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
python-dateutil==2.8.2
werkzeug==2.3.7
numpy==1.24.3
black==23.3.0
flake8==6.0.0
pytest==7.4.0